refactor(userpoint): seperate userpoint logic to independent files
