+++ model/converter/user.go
@@ -4,7 +4,6 @@ import (
 	"jcourse_go/model/dto"
 	"jcourse_go/model/model"
 	"jcourse_go/model/po"
-	"jcourse_go/util"
 )
 
 func ConvertUserDetailFromPO(po po.UserPO) model.UserDetail {
@@ -47,12 +46,4 @@ func RemoveUserEmail(u *model.UserDetail, userID int64) {
 		return
 	}
 	u.Email = ""
-}
-
-func ConvertUserPointDetailItemFromPO(po po.UserPointDetailPO) model.UserPointDetailItem {
-	return model.UserPointDetailItem{
-		Time:        po.CreatedAt.Format(util.GoTimeLayout),
-		Value:       po.Value,
-		Description: po.Description,
-	}
 }
